refs #9 - CAS-login for redmine with public content.
With this hack, you can have public pages and CAS-authentication.
The only problem is that you can *only* have CAS-authentication.
So if you configure CAS incorrectly, or you don't have an
administrator with a CAS-account, you have to hack the database
to get access again.

I am looking for a fix.