refs #9 - CAS-login for redmine with public content.
This is my second attempt. If authentication is required, you will
be redirected to the CAS login page whenever you are not authenticated.
(Except when you visit the default login page.)

If authentication is not required, the login page will show a link,
that allows you to login using CAS.